This project is based on "HR Analytics: Job Change of Data Scientists, Predict who will move to a new job" Dataset from Kaggle. The whole data was divided to train and test set. The dataset is quite imbalanced. we had to take care of that. Most of the features are categorical, some with high cardinality. Here, at first we are doing all the necessary data preprocessing steps, model building, predictions, anslysis of results etc. in research environment (Jupyter notebook) interactively. Later we wrote production level code in python for the machine learning pipeline. Missing imputation is a part of our pipeline along with other necessary steps. We are using Scikit Learn Pipeline here for production code. We made sure that the predictions obtained in the Research Environment matched with the predictions that we got by wriring and running production level code of the machine learning pipeline.  In order to take care of the imbalanced training data, we are using resampling technique (hybrid resampling -> combination Over-Sampling (using SMOTE) and Under-Sampling (using RandomUnderSampler)). This improved results quite a lot.  We tried Random Forest and XGBoost model both and later selled with XGBoost model since it gave better predictions.
This project is based on "HR Analytics: Job Change of Data Scientists, Predict who will move to a new job" Dataset from Kaggle. The whole data was divided to train and test set. The dataset is quite imbalanced. we had to take care of that.
Most of the features are categorical, some with high cardinality. Here, at first we are doing all the necessary data preprocessing steps, model building, predictions, anslysis of results etc. in research environment (Jupyter notebook) interactively.
Later we wrote production level code in python for the machine learning pipeline. Missing imputation is a part of our pipeline along with other necessary steps. We are using Scikit Learn Pipeline here for production code. We made sure that the predictions obtained in the Research Environment matched with the predictions that we got by wriring and running production level code of the machine learning pipeline.
In order to take care of the imbalanced training data, we are using resampling technique (hybrid resampling -> combination Over-Sampling (using SMOTE) and Under-Sampling (using RandomUnderSampler)). This improved results quite a lot.
We tried Random Forest and XGBoost model both and later selled with XGBoost model since it gave better predictions.